+++ src/components/date_picker/super_date_picker/index.js
@@ -1,5 +1,3 @@
-import {
-  WrappedEuiSuperDatePicker,
+export {
+  EuiSuperDatePicker,
 } from './super_date_picker';
-
-export { WrappedEuiSuperDatePicker as EuiSuperDatePicker };